Cleaned up solution, added Ability Proxy class
Ability proxy generates callback functions to inject into instantiated AbilityProxy : Ability class
